. r move class to separate file
